Add Tox Metadata Verification Check.
This change adds a new Zuul Job/Check to verify the app
metadata.yaml file in this repo.  The check is performed
by a console script belonging to the sysinv/config repo.

The existing metadata.yaml used the at '@' symbol to
serve as a template for logic in the Rules file to find
and replace for debian packaging purposes. This needed
to be changed so the verification software could read
the file without throwing an error.

Test Plan:
PASS - Verify that the appropriate Zuul Job has run
       and passes.

Story: 2010929
Task: 49230

Change-Id: I0b53dda13ed92eee66f0e71d609fbc6b5e563bd6
Signed-off-by: Reed, Joshua <Joshua.Reed@windriver.com>