Fixing leases removal getting stuck during app removal
When performing an apply on the dell-storage configuration without
using the overrides, the removal process becomes stuck and does not
complete successfully.

The root cause is that the cleanup job csi-powerstore-clear-job remains
running. This job consists of removing leases created on application
apply command.

The job gets stuck because its ServiceAccount has already been deleted
during the removal process. To prevent this from happening, what we did
was to make all hook-weights of the resources the same (default),
so that everything is created and deleted together during the process,
avoiding the job getting stuck.

In addition, as an improvement, in some patches the kubectl version was
still 1.29. We bumped it to a more recent version so that everything is
standardized between all CSIs.

Test Plan:
  PASS: All packages built
  PASS: Upload/apply/remove/delete/update dell-storage app
  PASS: Check if all pods are running for all CSIs
  PASS: All MTC operations executed
  PASS: Apply without dell-storage overrides, abort, then the app was
        removed

Closes-bug: 2136736

Change-Id: Ibabfefe9b25c5a6b044a3849a6638f736c23b726
Signed-off-by: Gabriel Przybysz Gonçalves Júnior <gabriel.przybyszgoncalvesjunior@windriver.com>